all norminette problems fixed
