+++ Makefile
@@ -17,7 +17,8 @@ CFILES = main.c map.c\
 
 PARSER_FILES = 
 
-RAY_CASTING_CFILES = movement.c ray_casting.c miscellaneus.c drawing.c ray_vertical_lines.c
+RAY_CASTING_CFILES = movement.c ray_casting.c miscellaneus.c drawing.c ray_vertical_lines.c \
+ray_horizontal_lines.c
 
 # DIRECTORIES 
 SRC_DIR = src/
